+++ src/Http/Client/Cookies/CookiesManager.php
@@ -16,6 +16,7 @@ namespace Berlioz\Http\Client\Cookies;
 
 use ArrayIterator;
 use Berlioz\Http\Client\Exception\HttpClientException;
+use Berlioz\Http\Client\Exception\InvalidCookieDomainException;
 use Countable;
 use IteratorAggregate;
 use Psr\Http\Message\RequestInterface;
@@ -32,10 +33,16 @@ class CookiesManager implements IteratorAggregate, Countable
 
     /**
      * CookiesManager constructor.
+     *
+     * @param Cookie[] $cookies
+     *
+     * @throws HttpClientException
      */
-    public function __construct()
+    public function __construct(array $cookies = [])
     {
-        $this->cookies = [];
+        foreach ($cookies as $cookie) {
+            $this->addCookie($cookie);
+        }
     }
 
     /**
@@ -112,6 +119,8 @@ class CookiesManager implements IteratorAggregate, Countable
     /**
      * Add cookie.
      *
+     * Explicit application insertion; the caller is responsible for validating the cookie's source.
+     *
      * @param Cookie $cookie
      *
      * @return static
@@ -148,6 +157,8 @@ class CookiesManager implements IteratorAggregate, Countable
     /**
      * Add cookies from response
      *
+     * Cookies with a domain not authorized by the response URI are ignored individually.
+     *
      * @param UriInterface $uri
      * @param ResponseInterface $response
      *
@@ -159,7 +170,11 @@ class CookiesManager implements IteratorAggregate, Countable
         $cookies = $response->getHeader('Set-Cookie');
 
         foreach ($cookies as $raw) {
-            $this->addRawCookie($raw, $uri);
+            try {
+                $this->addRawCookie($raw, $uri);
+            } catch (InvalidCookieDomainException) {
+                continue;
+            }
         }
 
         return $this;